fix(ScheduleConfiguration): change age getter to return int

+++ src/main/java/com/praktyki/backend/business/value/ScheduleConfiguration.java
@@ -45,7 +45,7 @@ public class ScheduleConfiguration {
         return mCommissionRate;
     }
 
-    public double getAge() {
+    public int getAge() {
         return mAge;
     }
